use hw.ncpuonline in GetLogicalCPUCountUsingSysCtl on openbsd & netbsd
The number of Logical CPUs reported by the hw.ncpu sysctl is twice as high as
the actual number of CPUs running on OpenBSD >= 6.4. with hyperthreading
disabled (hw.smt=0). Try hw.cpuonline first and fall back to hw.ncpu
in case it fails (which is the case on older OpenBSD before 6.4).